requirements.txt
langchain
python-dotenv
ipykernel
langchain-community
pypdf
beautifulsoup4
arxiv
PyMuPDF
wikipedia
langchain-text-splitters
lxml
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit